+++ svelte_ast/test/check.dart
@@ -3,25 +3,24 @@
 import 'dart:convert';
 import 'dart:io';
 
+import 'package:logging/logging.dart';
 import 'package:stack_trace/stack_trace.dart';
 import 'package:svelte_ast/src/reflection.dart';
 import 'package:svelte_ast/svelte_ast.dart';
 
 const String string = '''
-<script context="module" >
-  const foo = 'bar';
-</script>
-
-<script>
-  foo;
-</script>''';
+<script generics="T extends Object"></script>''';
 
 void main() {
+  Logger logger = Logger('parse');
+  logger.onRecord.listen(print);
+
   try {
     Root root = parse(
       string,
       fileName: 'check.dart',
       uri: Uri.file('check.dart'),
+      logger: logger,
     );
 
     Map<String, Object?> json = root.toJson(mapper);